+++ core/security.py
@@ -1,0 +1,21 @@
+from django.contrib.auth.backends import RemoteUserBackend as dj_RemoteUserBackend
+from rest_framework import permissions
+
+class ObjectPermissions(permissions.DjangoObjectPermissions):
+    """
+    Similar to `DjangoObjectPermissions`, but adding 'view' permissions.
+    """
+    perms_map = {
+        'GET': ['%(app_label)s.view_%(model_name)s'],
+        'OPTIONS': ['%(app_label)s.view_%(model_name)s'],
+        'HEAD': ['%(app_label)s.view_%(model_name)s'],
+        'POST': ['%(app_label)s.add_%(model_name)s'],
+        'PUT': ['%(app_label)s.change_%(model_name)s'],
+        'PATCH': ['%(app_label)s.change_%(model_name)s'],
+        'DELETE': ['%(app_label)s.delete_%(model_name)s'],
+    }
+
+class RemoteUserBackend(dj_RemoteUserBackend):
+    def __init__(self, *args, **kwargs):
+        self.create_unknown_user = False
+
